modifié :         ROSI/rosi/registration/load.py

+++ ROSI/rosi/registration/load.py
@@ -92,7 +92,7 @@ def loadStack(fileImage : str,
           stmask = np.array(stmask,dtype=np.int64)
           print(stmask[np.where(stmask>0)])
           data = stmask.reshape(-1)
-          stmask = Nifti1Image(stmask,stack.affine,dtype=np.int64)
+          stmask = Nifti1Image(stmask,stack.affine)
           #data = stmask.get_fdata()
           #print(data[np.where(data>0)])
           print(data[data<1])
